+++ pystac/extensions/label.py
@@ -1,7 +1,4 @@
-"""Implements the Label extension.
-
-https://github.com/stac-extensions/label
-"""
+"""Implements the :stac-ext:`Label Extension <label>`."""
 
 from pystac.extensions.base import ExtensionManagementMixin, SummariesExtension
 from typing import Any, Dict, Iterable, List, Optional, Sequence, Union, cast
@@ -11,7 +8,7 @@ from pystac.serialization.identify import STACJSONDescription, STACVersionID
 from pystac.extensions.hooks import ExtensionHooks
 from pystac.utils import StringEnum, get_required, map_opt
 
-SCHEMA_URI = "https://stac-extensions.github.io/label/v1.0.0/schema.json"
+SCHEMA_URI = "https://stac-extensions.github.io/label/v1.0.1/schema.json"
 
 PREFIX = "label:"
 
@@ -66,6 +63,8 @@ class LabelClasses:
 
     Use :meth:`LabelClasses.create` to create a new instance from property values.
     """
+
+    properties: Dict[str, Any]
 
     def __init__(self, properties: Dict[str, Any]):
         self.properties = properties
@@ -150,6 +149,8 @@ class LabelCount:
     Use :meth:`LabelCount.create` to create a new instance.
     """
 
+    properties: Dict[str, Any]
+
     def __init__(self, properties: Dict[str, Any]):
         self.properties = properties
 
@@ -214,6 +215,8 @@ class LabelStatistics:
     Use :meth:`LabelStatistics.create` to create a new instance.
     """
 
+    properties: Dict[str, Any]
+
     def __init__(self, properties: Dict[str, Any]) -> None:
         self.properties = properties
 
@@ -277,6 +280,8 @@ class LabelOverview:
 
     Use :meth:`LabelOverview.create` to create a new instance.
     """
+
+    properties: Dict[str, Any]
 
     def __init__(self, properties: Dict[str, Any]):
         self.properties = properties
@@ -436,6 +441,9 @@ class LabelExtension(ExtensionManagementMixin[Union[pystac.Item, pystac.Collecti
        >>> item: pystac.Item = ...
        >>> label_ext = LabelExtension.ext(item)
     """
+
+    obj: pystac.Item
+    schema_uri: str
 
     def __init__(self, item: pystac.Item) -> None:
         self.obj = item
@@ -780,7 +788,10 @@ class SummariesLabelExtension(SummariesExtension):
 
 class LabelExtensionHooks(ExtensionHooks):
     schema_uri: str = SCHEMA_URI
-    prev_extension_ids = {"label"}
+    prev_extension_ids = {
+        "label",
+        "https://stac-extensions.github.io/label/v1.0.0/schema.json",
+    }
     stac_object_types = {pystac.STACObjectType.ITEM}
 
     def get_object_links(