+++ pystac/extensions/table.py
@@ -1,7 +1,4 @@
-"""Implements the Table extension
-
-https://github.com/stac-extensions/table
-"""
+"""Implements the :stac-ext:`Table Extension <table>.`"""
 from typing import Any, Dict, Generic, List, Optional, TypeVar, Union, cast
 
 import pystac
@@ -183,7 +180,7 @@ class TableExtension(
     @property
     def primary_geometry(self) -> Optional[str]:
         """The primary geometry column name"""
-        return self.properties.get(PRIMARY_GEOMETRY_PROP)
+        return self._get_property(PRIMARY_GEOMETRY_PROP, str)
 
     @primary_geometry.setter
     def primary_geometry(self, v: Optional[str]) -> None:
@@ -195,7 +192,7 @@ class TableExtension(
     @property
     def row_count(self) -> Optional[int]:
         """The number of rows in the dataset"""
-        return self.properties.get(ROW_COUNT_PROP)
+        return self._get_property(ROW_COUNT_PROP, int)
 
     @row_count.setter
     def row_count(self, v: Optional[int]) -> None: